requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
regex